requirements.txt
pandas
pinecone-client
openai
python-dotenv
numpy 